@@ -4,18 +4,20 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=3.4.0 <4.0.0"
+  sdk: ">=3.9.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   # Utility
-  uuid: ^4.4.2
-  intl: ^0.19.0
+  uuid: ^4.5.1
+  intl: ^0.20.2
 
   # UI
-  flutter_credit_card: ^4.0.1
+  flutter_credit_card: ^4.1.0
+
+  webview_flutter: ^4.13.0
 
   liqpay:
     path: ../
@@ -23,7 +25,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^4.0.0
+  flutter_lints: ^6.0.0
 
 flutter:
   uses-material-design: true
