@@ -1,4 +1,4 @@
-// Mocks generated by Mockito 5.4.4 from annotations
+// Mocks generated by Mockito 5.4.6 from annotations
 // in liqpay/test/liqpay_test.dart.
 // Do not manually edit this file.
 
@@ -19,10 +19,12 @@ import 'package:mockito/src/dummies.dart' as _i5;
 // ignore_for_file: deprecated_member_use_from_same_package
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
+// ignore_for_file: must_be_immutable
 // ignore_for_file: prefer_const_constructors
 // ignore_for_file: unnecessary_parenthesis
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
+// ignore_for_file: invalid_use_of_internal_member
 
 class _FakeResponse_0 extends _i1.SmartFake implements _i2.Response {
   _FakeResponse_0(Object parent, Invocation parentInvocation)
@@ -55,6 +57,7 @@ class MockClient extends _i1.Mock implements _i2.Client {
             ),
           )
           as _i3.Future<_i2.Response>);
+
   @override
   _i3.Future<_i2.Response> get(Uri? url, {Map<String, String>? headers}) =>
       (super.noSuchMethod(
@@ -67,6 +70,7 @@ class MockClient extends _i1.Mock implements _i2.Client {
             ),
           )
           as _i3.Future<_i2.Response>);
+
   @override
   _i3.Future<_i2.Response> post(
     Uri? url, {
@@ -92,6 +96,7 @@ class MockClient extends _i1.Mock implements _i2.Client {
             ),
           )
           as _i3.Future<_i2.Response>);
+
   @override
   _i3.Future<_i2.Response> put(
     Uri? url, {
@@ -117,6 +122,7 @@ class MockClient extends _i1.Mock implements _i2.Client {
             ),
           )
           as _i3.Future<_i2.Response>);
+
   @override
   _i3.Future<_i2.Response> patch(
     Uri? url, {
@@ -142,6 +148,7 @@ class MockClient extends _i1.Mock implements _i2.Client {
             ),
           )
           as _i3.Future<_i2.Response>);
+
   @override
   _i3.Future<_i2.Response> delete(
     Uri? url, {
@@ -167,6 +174,7 @@ class MockClient extends _i1.Mock implements _i2.Client {
             ),
           )
           as _i3.Future<_i2.Response>);
+
   @override
   _i3.Future<String> read(Uri? url, {Map<String, String>? headers}) =>
       (super.noSuchMethod(
@@ -179,6 +187,7 @@ class MockClient extends _i1.Mock implements _i2.Client {
             ),
           )
           as _i3.Future<String>);
+
   @override
   _i3.Future<_i6.Uint8List> readBytes(
     Uri? url, {
@@ -189,6 +198,7 @@ class MockClient extends _i1.Mock implements _i2.Client {
             returnValue: _i3.Future<_i6.Uint8List>.value(_i6.Uint8List(0)),
           )
           as _i3.Future<_i6.Uint8List>);
+
   @override
   _i3.Future<_i2.StreamedResponse> send(_i2.BaseRequest? request) =>
       (super.noSuchMethod(
@@ -201,6 +211,7 @@ class MockClient extends _i1.Mock implements _i2.Client {
             ),
           )
           as _i3.Future<_i2.StreamedResponse>);
+
   @override
   void close() => super.noSuchMethod(
     Invocation.method(#close, []),
